+++ classes/Supplier.php
@@ -186,21 +186,23 @@ class SupplierCore extends ObjectModel
 
 	public static function getProducts($id_supplier, $id_lang, $p, $n, $orderBy = NULL, $orderWay = NULL, $getTotal = false, $active = true, $active_category = true)
 	{
-		if ($p < 1) $p = 1;
-	 	if (empty($orderBy) OR $orderBy == 'position') $orderBy = 'name';
-	 	if (empty($orderWay)) $orderWay = 'ASC';
+		if ($p < 1)
+			$p = 1;
+	 	if (empty($orderBy) || $orderBy == 'position')
+			$orderBy = 'name';
+	 	if (empty($orderWay))
+			$orderWay = 'ASC';
 
-		if (!Validate::isOrderBy($orderBy) OR !Validate::isOrderWay($orderWay))
+		if (!Validate::isOrderBy($orderBy) || !Validate::isOrderWay($orderWay))
 			die (Tools::displayError());
-			
+
 		$groups = FrontController::getCurrentCustomerGroups();
 		$sqlGroups = (count($groups) ? 'IN ('.implode(',', $groups).')' : '= 1');
 
 		/* Return only the number of products */
 		if ($getTotal)
-		{
-			$sql = '
-				SELECT p.`id_product`
+			return (int)Db::getInstance(_PS_USE_SQL_SLAVE_)->getValue('
+				SELECT count(p.`id_product`)
 				FROM `'._DB_PREFIX_.'product` p
 				WHERE p.id_supplier = '.(int)($id_supplier)
 				.($active ? ' AND p.`active` = 1' : '').'
@@ -210,26 +212,23 @@ class SupplierCore extends ObjectModel
 					LEFT JOIN `'._DB_PREFIX_.'category_product` cp ON (cp.`id_category` = cg.`id_category`)'.
 					($active_category ? ' INNER JOIN `'._DB_PREFIX_.'category` ca ON cp.`id_category` = ca.`id_category` AND ca.`active` = 1' : '').'
 					WHERE cg.`id_group` '.$sqlGroups.'
-				)';
-			$result = Db::getInstance(_PS_USE_SQL_SLAVE_)->ExecuteS($sql);
-			return (int)(sizeof($result));
-		}
+				)');
 
 		$sql = '
 		SELECT p.*, pl.`description`, pl.`description_short`, pl.`link_rewrite`, pl.`meta_description`, pl.`meta_keywords`, pl.`meta_title`, pl.`name`, i.`id_image`, il.`legend`, s.`name` AS supplier_name, tl.`name` AS tax_name, t.`rate`, DATEDIFF(p.`date_add`, DATE_SUB(NOW(), INTERVAL '.(Validate::isUnsignedInt(Configuration::get('PS_NB_DAYS_NEW_PRODUCT')) ? Configuration::get('PS_NB_DAYS_NEW_PRODUCT') : 20).' DAY)) > 0 AS new,
 			(p.`price` * ((100 + (t.`rate`))/100)) AS orderprice, m.`name` AS manufacturer_name
 		FROM `'._DB_PREFIX_.'product` p
-		LEFT JOIN `'._DB_PREFIX_.'product_lang` pl ON (p.`id_product` = pl.`id_product` AND pl.`id_lang` = '.(int)($id_lang).')
+		LEFT JOIN `'._DB_PREFIX_.'product_lang` pl ON (p.`id_product` = pl.`id_product` AND pl.`id_lang` = '.(int)$id_lang.')
 		LEFT JOIN `'._DB_PREFIX_.'image` i ON (i.`id_product` = p.`id_product` AND i.`cover` = 1)
-		LEFT JOIN `'._DB_PREFIX_.'image_lang` il ON (i.`id_image` = il.`id_image` AND il.`id_lang` = '.(int)($id_lang).')
+		LEFT JOIN `'._DB_PREFIX_.'image_lang` il ON (i.`id_image` = il.`id_image` AND il.`id_lang` = '.(int)$id_lang.')
 		LEFT JOIN `'._DB_PREFIX_.'tax_rule` tr ON (p.`id_tax_rules_group` = tr.`id_tax_rules_group`
 		                                           AND tr.`id_country` = '.(int)Country::getDefaultCountryId().'
 	                                           	   AND tr.`id_state` = 0)
 	    LEFT JOIN `'._DB_PREFIX_.'tax` t ON (t.`id_tax` = tr.`id_tax`)
-		LEFT JOIN `'._DB_PREFIX_.'tax_lang` tl ON (t.`id_tax` = tl.`id_tax` AND tl.`id_lang` = '.(int)($id_lang).')
+		LEFT JOIN `'._DB_PREFIX_.'tax_lang` tl ON (t.`id_tax` = tl.`id_tax` AND tl.`id_lang` = '.(int)$id_lang.')
 		LEFT JOIN `'._DB_PREFIX_.'supplier` s ON s.`id_supplier` = p.`id_supplier`
 		LEFT JOIN `'._DB_PREFIX_.'manufacturer` m ON m.`id_manufacturer` = p.`id_manufacturer`
-		WHERE p.`id_supplier` = '.(int)($id_supplier).($active ? ' AND p.`active` = 1' : '').'
+		WHERE p.`id_supplier` = '.(int)$id_supplier.($active ? ' AND p.`active` = 1' : '').'
 		AND p.`id_product` IN (
 					SELECT cp.`id_product`
 					FROM `'._DB_PREFIX_.'category_group` cg
@@ -256,6 +255,7 @@ class SupplierCore extends ObjectModel
 		LEFT JOIN `'._DB_PREFIX_.'product_lang` pl ON (p.`id_product` = pl.`id_product` AND pl.`id_lang` = '.(int)($id_lang).')
 		WHERE p.`id_supplier` = '.(int)($this->id));
 	}
+
 	/*
 	* Specify if a supplier already in base
 	*
@@ -271,11 +271,12 @@ class SupplierCore extends ObjectModel
 
 		return isset($row['id_supplier']);
 	}
-	
+
 	public function delete()
 	{
-		if (parent::delete())
+		if (Db::getInstance()->Execute('UPDATE `'._DB_PREFIX_.'product` SET `id_supplier` = 0 WHERE `id_supplier` = '.(int)$this->id) && parent::delete())
 			return $this->deleteImage();
+		return false;
 	}
 }
